feat: move remaining pure collection utils into @mulmoclaude/collection-plugin (chunk 1b-rest)
Move the last three framework-free, schema-typed utilities — sortItems,
itemLabel, calendarGrid — into the package core, now that the canonical schema
types live there (chunk 1b). sortItems' type refs retargeted to the package's
own ./schema using canonical names (CollectionFieldSpec/CollectionFieldType).

Rewired importers: CollectionView / CollectionDayView / CollectionCalendarView,
the collectionViewMode sibling (SortState), and 3 unit tests. The frontend-only
utils (draft / enumColors / collectionViewMode) and host-coupled notifiedItems
stay in src/ for the later Vue chunk.

Verified: yarn typecheck (server/vue/test/packages), lint, build all green;
59 collection util tests pass.
