+++ packages/plugins/collection-plugin/src/index.ts
@@ -9,3 +9,6 @@ export * from "./core/schema";
 export * from "./core/actionVisible";
 export * from "./core/derivedFormula";
 export * from "./core/deriveAll";
+export * from "./core/sortItems";
+export * from "./core/itemLabel";
+export * from "./core/calendarGrid";
