+++ src/components/CollectionCalendarView.vue
@@ -96,9 +96,9 @@
 <script setup lang="ts">
 import { computed, ref } from "vue";
 import { useI18n } from "vue-i18n";
-import { bucketRecords, buildMonthGrid, ymdKey, daySlice, MINUTES_PER_DAY, type Ymd, type RecordSpan, type DaySlice } from "../utils/collections/calendarGrid";
+import { bucketRecords, buildMonthGrid, ymdKey, daySlice, MINUTES_PER_DAY, type Ymd, type RecordSpan, type DaySlice } from "@mulmoclaude/collection-plugin";
 import { resolveEnumColor, type EnumColorClasses } from "../utils/collections/enumColors";
-import { labelFieldFor, itemIdOf, itemLabelOf } from "../utils/collections/itemLabel";
+import { labelFieldFor, itemIdOf, itemLabelOf } from "@mulmoclaude/collection-plugin";
 import type { CollectionItem, CollectionSchema } from "./collectionTypes";
 
 const props = defineProps<{
