+++ src/components/CollectionView.vue
@@ -733,7 +733,20 @@ import CollectionCalendarView from "./CollectionCalendarView.vue";
 import CollectionDayView from "./CollectionDayView.vue";
 import CollectionKanbanView from "./CollectionKanbanView.vue";
 import CollectionCustomView from "./CollectionCustomView.vue";
-import { dateOf, type Ymd } from "../utils/collections/calendarGrid";
+import {
+  dateOf,
+  type Ymd,
+  isSortableField,
+  nextSortDirection,
+  sortItems,
+  numericSortValue,
+  stringSortValue,
+  dateSortValue,
+  enumSortValue,
+  boolSortValue,
+  type SortState,
+  type SortValue,
+} from "@mulmoclaude/collection-plugin";
 import { useConfirm } from "../composables/useConfirm";
 import { useAppApi } from "../composables/useAppApi";
 import { useShortcuts } from "../composables/useShortcuts";
@@ -747,18 +760,6 @@ import {
   type CollectionViewMode,
   type BuiltInViewMode,
 } from "../utils/collections/collectionViewMode";
-import {
-  isSortableField,
-  nextSortDirection,
-  sortItems,
-  numericSortValue,
-  stringSortValue,
-  dateSortValue,
-  enumSortValue,
-  boolSortValue,
-  type SortState,
-  type SortValue,
-} from "../utils/collections/sortItems";
 import { useCollectionRendering } from "../composables/collections/useCollectionRendering";
 import { buildUpdatedRecord, coerceInlineValue, draftToRecord, firstMissingRequiredField, rowFromItem } from "../utils/collections/draft";
 import type {
