+++ src/utils/collections/collectionViewMode.ts
@@ -5,7 +5,7 @@
 // own copy in the tool-result `viewState` (they read, never write — so a
 // stale card can't clobber the shared preference).
 
-import type { SortState } from "./sortItems";
+import type { SortState } from "@mulmoclaude/collection-plugin";
 
 /** The host's built-in, field-derived view modes. */
 export type BuiltInViewMode = "table" | "calendar" | "kanban";
