+++ test/utils/collections/test_sortItems.ts
@@ -14,7 +14,7 @@ import {
   enumSortValue,
   boolSortValue,
   type SortValue,
-} from "../../../src/utils/collections/sortItems.js";
+} from "@mulmoclaude/collection-plugin";
 import type { CollectionItem } from "../../../src/components/collectionTypes.js";
 
 describe("isSortableField", () => {
